Move ChapterComparator into chapter package
